requirements.txt
amqp==1.4.9
anyjson==0.3.3
appnope==0.1.0
asn1crypto==0.24.0
Babel==2.6.0
backcall==0.1.0
billiard==3.3.0.23
celery==3.1.26.post2
celery-with-redis==3.0
cffi==1.11.5
colorama==0.3.9
cryptography==2.3.1
decorator==4.3.0
diff-match-patch==20121119
Django==1.11
django-celery==3.2.2
django-crispy-forms==1.7.2
django-debug-toolbar==1.9.1
django-formtools==2.1
django-haystack==2.8.1
django-import-export==1.0.1
django-redis-sessions==0.6.1
django-reversion==3.0.0
djangorestframework==3.8.2
DjangoUeditor==1.8.143
et-xmlfile==1.0.1
flower==0.9.2
future==0.16.0
httplib2==0.9.2
idna==2.7
ipython==6.4.0
ipython-genutils==0.2.0
jdcal==1.4
jedi==0.12.0
kombu==3.0.37
Markdown==2.6.11
mysqlclient==1.3.12
odfpy==1.3.6
openpyxl==2.5.5
parso==0.2.1
pexpect==4.6.0
pickleshare==0.7.4
Pillow==5.2.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
pycparser==2.18
Pygments==2.2.0
PyMySQL==0.9.2
pytz==2018.4
PyYAML==3.13
redis==2.10.6
simplegeneric==0.8.1
six==1.11.0
sqlparse==0.2.4
tablib==0.12.1
tornado==5.1
traitlets==4.3.2
unicodecsv==0.14.1
wcwidth==0.1.7
Whoosh==2.7.4
xlrd==1.1.0
xlwt==1.3.0
